run_qemu.sh: default to build OS like mkosi instead of Fedora 40
Using one distro to build another is fraught with peril and most people
just want to stick to the same OS. It's annoying and time-consuming not
to forget to tell run_qemu.sh what distribution it is currently on.
That information is available in /etc/os-release so let's just get it
and use it.

Do keep things simple and transparent, pass $distro and $rev user input
"as is" to mkosi configuration; including missing or empty parameters
which default to the same OS and same release.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>